AlarmManager && Notification

+++ app/src/main/java/com/todo/fursa/room/dao/TodoDAO.kt
@@ -23,13 +23,13 @@ interface TodoDAO {
     fun save(todo: Todo?)
 
     @Query("DELETE FROM todo WHERE id=:todoId")
-    fun delete(todoId: Int)
+    fun delete(todoId: Long)
 
     @Query("DELETE FROM todo")
     fun deleteAll()
 
     @Query("SELECT * FROM todo WHERE id=:todoId")
-    fun selectById(todoId: Int): LiveData<Todo>
+    fun selectById(todoId: Long): LiveData<Todo>
 
     @Query("SELECT COUNT(id) FROM todo")
     fun size(): LiveData<Int>
